+++ VComponents/Components/Item Pickers/Wheel Picker/VWheelPicker.swift
@@ -14,11 +14,11 @@ import SwiftUI
 ///
 /// Best suited for `5`+ items.
 ///
-/// Model, state, header, and footer can be passed as parameters.
+/// Model, header, and footer can be passed as parameters.
 ///
 /// Usage example:
 ///
-///     @State var selectedIndex: Int = 7
+///     @State var selection: String = "January"
 ///
 ///     let rowTitles: [String] = [
 ///         "January", "February", "March",
@@ -29,7 +29,7 @@ import SwiftUI
 ///
 ///     var body: some View {
 ///         VWheelPicker(
-///             selectedIndex: $selectedIndex,
+///             selection: $selection,
 ///             headerTitle: "Lorem ipsum dolor sit amet",
 ///             footerTitle: "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
 ///             rowTitles: rowTitles
@@ -46,23 +46,22 @@ public struct VWheelPicker<Data, RowContent>: View
     // MARK: Properties
     private let model: VWheelPickerModel
     
-    private let state: VWheelPickerState
+    @Environment(\.isEnabled) private var isEnabled: Bool
+    private var internalState: VWheelPickerInternalState { .init(isEnabled: isEnabled) }
     
     @Binding private var selectedIndex: Int
     
     private let headerTitle: String?
     private let footerTitle: String?
     
-    private let data: Data
-    private let rowContent: (Data.Element) -> RowContent
+    private let content: VWheelPickerContent<Data, RowContent>
     
     @State private var rowWidth: CGFloat = .zero
     
-    // MARK: Initializers - View Builder
-    /// Initializes component with selected index, header, footer, data, and row content.
+    // MARK: Initializers - Index
+    /// Initializes component with selected index, data, and row content.
     public init(
         model: VWheelPickerModel = .init(),
-        state: VWheelPickerState = .enabled,
         selectedIndex: Binding<Int>,
         headerTitle: String? = nil,
         footerTitle: String? = nil,
@@ -70,50 +69,81 @@ public struct VWheelPicker<Data, RowContent>: View
         @ViewBuilder rowContent: @escaping (Data.Element) -> RowContent
     ) {
         self.model = model
-        self.state = state
         self._selectedIndex = selectedIndex
         self.headerTitle = headerTitle
         self.footerTitle = footerTitle
-        self.data = data
-        self.rowContent = rowContent
+        self.content = .custom(data: data, rowContent: rowContent)
     }
 
-    // MARK: Initializes: Row Titles
-    /// Initializes component with selected index, header, footer, and row titles.
+    /// Initializes component with selected index and row titles.
     public init(
         model: VWheelPickerModel = .init(),
-        state: VWheelPickerState = .enabled,
         selectedIndex: Binding<Int>,
         headerTitle: String? = nil,
         footerTitle: String? = nil,
         rowTitles: [String]
     )
         where
-            Data == Array<String>,
-            RowContent == VText
+            Data == Array<Never>,
+            RowContent == Never
     {
-        self.init(
-            model: model,
-            state: state,
-            selectedIndex: selectedIndex,
-            headerTitle: headerTitle,
-            footerTitle: footerTitle,
-            data: rowTitles,
-            rowContent: { title in
-                VText(
-                    color: model.colors.textContent.for(state),
-                    font: model.fonts.rows,
-                    title: title
-                )
-            }
-        )
+        self.model = model
+        self._selectedIndex = selectedIndex
+        self.headerTitle = headerTitle
+        self.footerTitle = footerTitle
+        self.content = .titles(titles: rowTitles)
     }
-
-    // MARK: Initialzers: Pickable Item
-    /// Initializes component with `PickableEnumeration`, header, footer, and row content.
+    
+    // MARK: Initializers - Hashable
+    /// Initializes component with selection value, data, and row content.
+    public init<SelectionValue>(
+        model: VWheelPickerModel = .init(),
+        selection: Binding<SelectionValue>,
+        headerTitle: String? = nil,
+        footerTitle: String? = nil,
+        data: Data,
+        @ViewBuilder rowContent: @escaping (Data.Element) -> RowContent
+    )
+        where
+            Data == Array<SelectionValue>,
+            SelectionValue: Hashable
+    {
+        self.model = model
+        self._selectedIndex = .init(
+            get: { data.firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = data[$0] }
+        )
+        self.headerTitle = headerTitle
+        self.footerTitle = footerTitle
+        self.content = .custom(data: data, rowContent: rowContent)
+    }
+    
+    /// Initializes component with selection value and row titles.
+    public init(
+        model: VWheelPickerModel = .init(),
+        selection: Binding<String>,
+        headerTitle: String? = nil,
+        footerTitle: String? = nil,
+        rowTitles: [String]
+    )
+        where
+            Data == Array<Never>,
+            RowContent == Never
+    {
+        self.model = model
+        self._selectedIndex = .init(
+            get: { rowTitles.firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = rowTitles[$0] }
+        )
+        self.headerTitle = headerTitle
+        self.footerTitle = footerTitle
+        self.content = .titles(titles: rowTitles)
+    }
+    
+    // MARK: Initializers - Pickable Enumeration & Pickable Titled Enumeration
+    /// Initializes component with `PickableEnumeration` and row content.
     public init<PickableItem>(
         model: VWheelPickerModel = .init(),
-        state: VWheelPickerState = .enabled,
         selection: Binding<PickableItem>,
         headerTitle: String? = nil,
         footerTitle: String? = nil,
@@ -123,108 +153,106 @@ public struct VWheelPicker<Data, RowContent>: View
             Data == Array<PickableItem>,
             PickableItem: PickableEnumeration
     {
-        self.init(
-            model: model,
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            headerTitle: headerTitle,
-            footerTitle: footerTitle,
-            data: .init(PickableItem.allCases),
-            rowContent: rowContent
+        self.model = model
+        self._selectedIndex = .init(
+            get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
         )
+        self.headerTitle = headerTitle
+        self.footerTitle = footerTitle
+        self.content = .custom(data: Array(PickableItem.allCases), rowContent: rowContent)
     }
-
-    // MARK: Initialzers: Pickable Titled Item
-    /// Initializes component with `PickableTitledItem`, header, and footer.
+    
+    /// Initializes component with `PickableTitledEnumeration`.
     public init<PickableItem>(
         model: VWheelPickerModel = .init(),
-        state: VWheelPickerState = .enabled,
         selection: Binding<PickableItem>,
         headerTitle: String? = nil,
         footerTitle: String? = nil
     )
         where
-            Data == Array<PickableItem>,
-            RowContent == VText,
+            Data == Array<Never>,
+            RowContent == Never,
             PickableItem: PickableTitledEnumeration
     {
-        self.init(
-            model: model,
-            state: state,
-            selectedIndex: .init(
-                get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
-                set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
-            ),
-            headerTitle: headerTitle,
-            footerTitle: footerTitle,
-            data: .init(PickableItem.allCases),
-            rowContent: { item in
-                VText(
-                    color: model.colors.textContent.for(state),
-                    font: model.fonts.rows,
-                    title: item.pickerTitle
-                )
-            }
+        self.model = model
+        self._selectedIndex = .init(
+            get: { Array(PickableItem.allCases).firstIndex(of: selection.wrappedValue)! }, // fatalError
+            set: { selection.wrappedValue = Array(PickableItem.allCases)[$0] }
         )
+        self.headerTitle = headerTitle
+        self.footerTitle = footerTitle
+        self.content = .titles(titles: Array(PickableItem.allCases).map { $0.pickerTitle })
     }
 
     // MARK: Body
     public var body: some View {
         VStack(alignment: .leading, spacing: model.layout.headerFooterSpacing, content: {
-            headerView
-            pickerView
-            footerView
+            header
+            picker
+            footer
         })
     }
     
-    private var pickerView: some View {
-        Picker(selection: $selectedIndex, label: EmptyView(), content: {
-            ForEach(0..<data.count, content: { i in
-                rowContent(data[i])
-                    .tag(i)
-            })
-        })
-            .pickerStyle(.wheel)
-            .labelsHidden()
-            
-            .disabled(!state.isEnabled) // Luckily, doesn't affect colors
-            .opacity(model.colors.content.for(state))
-            
-            .background(model.colors.background.for(state).cornerRadius(model.layout.cornerRadius))
-    }
-    
-    @ViewBuilder private var headerView: some View {
+    @ViewBuilder private var header: some View {
         if let headerTitle = headerTitle, !headerTitle.isEmpty {
             VText(
-                color: model.colors.header.for(state),
+                color: model.colors.header.for(internalState),
                 font: model.fonts.header,
                 title: headerTitle
             )
                 .padding(.horizontal, model.layout.headerMarginHorizontal)
-                .opacity(model.colors.content.for(state))
         }
     }
     
-    @ViewBuilder private var footerView: some View {
+    @ViewBuilder private var footer: some View {
         if let footerTitle = footerTitle, !footerTitle.isEmpty {
             VText(
                 type: .multiLine(alignment: .leading, limit: nil),
-                color: model.colors.footer.for(state),
+                color: model.colors.footer.for(internalState),
                 font: model.fonts.footer,
                 title: footerTitle
             )
                 .padding(.horizontal, model.layout.headerMarginHorizontal)
-                .opacity(model.colors.content.for(state))
+        }
+    }
+    
+    private var picker: some View {
+        Picker(
+            selection: $selectedIndex,
+            content: rows,
+            label: { EmptyView() }
+        )
+            .pickerStyle(.wheel)
+            .disabled(!internalState.isEnabled) // Luckily, doesn't affect colors
+            .background(model.colors.background.for(internalState).cornerRadius(model.layout.cornerRadius))
+    }
+    
+    @ViewBuilder private func rows() -> some View {
+        switch content {
+        case .titles(let titles):
+            ForEach(0..<titles.count, id: \.self, content: { i in
+                VText(
+                    color: model.colors.title.for(internalState),
+                    font: model.fonts.rows,
+                    title: titles[i]
+                )
+                    .tag(i)
+            })
+            
+        case .custom(let data, let rowContent):
+            ForEach(0..<data.count, id: \.self, content: { i in
+                rowContent(data[i])
+                    .opacity(model.colors.customContentOpacities.for(internalState))
+                    .tag(i)
+            })
         }
     }
 }
 
 // MARK: - Preview
 struct VWheelPicker_Previews: PreviewProvider {
-    @State private static var selectedIndex: Int = 7
+    @State private static var selection: String = rowTitles.first!
     
     private static var rowTitles: [String] {
         [
@@ -237,7 +265,7 @@ struct VWheelPicker_Previews: PreviewProvider {
 
     static var previews: some View {
         VWheelPicker(
-            selectedIndex: $selectedIndex,
+            selection: $selection,
             headerTitle: "Lorem ipsum dolor sit amet",
             footerTitle: "Lorem ipsum dolor sit amet, consectetur adipiscing elit",
             rowTitles: rowTitles