+++ VComponents/Components/Item Pickers/Wheel Picker/VWheelPickerContent.swift
@@ -1,0 +1,11 @@
+//
+//  VWheelPickerContent.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 3/7/22.
+//
+
+import Foundation
+
+// MARK: - V Wheel Picker Content
+typealias VWheelPickerContent = VSegmentedPickerContent