+++ VComponentsDemo/Scenes/Demos/Item Pickers/VWheelPickerDemoView.swift
@@ -13,20 +13,11 @@ struct VWheelPickerDemoView: View {
     // MARK: Properties
     static var navBarTitle: String { "Wheel Picker" }
     
-    @State private var selection: VSegmentedPickerDataSource = .green
-    @State private var state: VWheelPickerState = .enabled
-    @State private var labelType: VSegmentedPickerContent = .title
+    @State private var selection: VWheelPickerDataSource = .green
+    @State private var isEnabled: Bool = true
+    @State private var contentType: VWheelPickerContent = .title
     @State private var hasHeader: Bool = true
     @State private var hasFooter: Bool = true
-    @State private var loweredOpacityWhenDisabled: Bool = VWheelPickerModel.Colors().content.disabledOpacity != 1
-
-    private var model: VWheelPickerModel {
-        var model: VWheelPickerModel = .init()
-        
-        model.colors.content.disabledOpacity = loweredOpacityWhenDisabled ? 0.5 : 1
-        
-        return model
-    }
 
     // MARK: Body
     var body: some View {
@@ -34,60 +25,57 @@ struct VWheelPickerDemoView: View {
             .standardNavigationTitle(Self.navBarTitle)
     }
     
-    @ViewBuilder private func component() -> some View {
-        switch labelType {
-        case .title:
-            VWheelPicker(
-                model: model,
-                state: state,
-                selection: $selection,
-                headerTitle: hasHeader ? "Lorem ipsum dolor sit amet" : nil,
-                footerTitle: hasFooter ? "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam tincidunt ante at finibus cursus." : nil
-            )
-        
-        case .custom:
-            VWheelPicker(
-                model: model,
-                state: state,
-                selection: $selection,
-                headerTitle: hasHeader ? "Lorem ipsum dolor sit amet" : nil,
-                footerTitle: hasFooter ? "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam tincidunt ante at finibus cursus." : nil,
-                rowContent: { $0.pickerSymbol }
-            )
-        }
+    private func component() -> some View {
+        Group(content: {
+            switch contentType {
+            case .title:
+                VWheelPicker(
+                    selection: $selection,
+                    headerTitle: hasHeader ? "Lorem ipsum dolor sit amet" : nil,
+                    footerTitle: hasFooter ? "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam tincidunt ante at finibus cursus." : nil
+                )
+
+            case .custom:
+                VWheelPicker(
+                    selection: $selection,
+                    headerTitle: hasHeader ? "Lorem ipsum dolor sit amet" : nil,
+                    footerTitle: hasFooter ? "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nullam tincidunt ante at finibus cursus." : nil,
+                    rowContent: { $0.pickerSymbol }
+                )
+            }
+        })
+            .disabled(!isEnabled)
     }
     
     @DemoViewSettingsSectionBuilder private func settings() -> some View {
         DemoViewSettingsSection(content: {
-            VSegmentedPicker(selection: $state, headerTitle: "State")
+            VSegmentedPicker(
+                selection: .init(
+                    get: { VWheelPickerState(isEnabled: isEnabled) },
+                    set: { isEnabled = $0 == .enabled }
+                ),
+                headerTitle: "State"
+            )
         })
-        
+
         DemoViewSettingsSection(content: {
-            VSegmentedPicker(selection: $labelType, headerTitle: "Content")
+            VSegmentedPicker(selection: $contentType, headerTitle: "Content")
         })
-        
+
         DemoViewSettingsSection(content: {
             ToggleSettingView(isOn: $hasHeader, title: "Header")
-            
+
             ToggleSettingView(isOn: $hasFooter, title: "Footer")
-        })
-        
-        DemoViewSettingsSection(content: {
-            ToggleSettingView(isOn: $loweredOpacityWhenDisabled, title: "Low Disabled Opacity", description: "Content lowers opacity when disabled")
         })
     }
 }
 
 // MARK: - Helpers
-extension VWheelPickerState: PickableTitledEnumeration {
-    public var pickerTitle: String {
-        switch self {
-        case .enabled: return "Enabled"
-        case .disabled: return "Disabled"
-        @unknown default: fatalError()
-        }
-    }
-}
+private typealias VWheelPickerState = VSecondaryButtonState
+
+private typealias VWheelPickerContent = VSegmentedPickerContent
+
+private typealias VWheelPickerDataSource = VSegmentedPickerDataSource
 
 // MARK: - Preview
 struct VWheelPickerDemoView_Previews: PreviewProvider {